IOT-1072 Unfork libcoap: update forked headers
Make the .h files compatible between the forked libcoap and the official
libcoap, so that no breaking changes (only additions) are done to the
official libcoap, which other projects depend on the current
types/prototypes in.

Make any additions (that are beyond the official libcoap, but that are
already used by IoTivity code) to libcoap .h files follow the conventions of
libcoap, to make it more likely they'll be accepted back into the upstream
libcoap.  E.g., coap_transport_type -> coap_transport_t, etc.

Leaving WITH_UPSTREAM_LIBCOAP off by default on all OS's in this change.
Turning it on by default in certain OS's will be done as part of a
separate change.

Change-Id: I373ef818b8aeb7608c5e3511dc9880047946fbe0
Signed-off-by: Dave Thaler <dthaler@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9373
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: David Antler <david.a.antler@intel.com>