Add sample provisioning servers in package
Added sample provisioning servers into the test package built
Corresponding .dat files are also added to same location as binary file
,i.e. /usr/bin/provision-sample
Server opens the dat file from the current location of the binary file.
This needs to be updated for tizen, to be in a place out of bin, where Read-Write
is possible.

Change-Id: I0d13386976d2d4910934064d6092e1b360dcf919
Credit-to: Ashwini Kumar <k.ashwini@samsung.com>
Signed-off-by: Ashwini Kumar <k.ashwini@samsung.com>
Signed-off-by: Philippe Coval <philippe.coval@osg.samsung.com>
Signed-off-by: Ashwini Kumar <k.ashwini@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9311
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mushfiqul Islam <i.mushfiq@samsung.com>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>
Signed-off-by: Ashwini Kumar <k.ashwini@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9609
Reviewed-by: Ziran Sun <ziran.sun@samsung.com>