Added API for setting representation in RCS Client of Resource Encapsulation.
RCSRemoteResourceObject of RE layer provides an API to get remote resource
representation using RCSRepresentation Object, but it is missing the API to
set RCSRepresentation.

Setting of RCSRepresentation is required for collection resources.

In this patch I have added the set API which accepts RCSRepresentation as input
parameter so it can be used to send POST request on collection resources where
payload contains attributes to be set on multiple child resources.

Signed-off-by: Abhishek Pandey <abhi.siso@samsung.com>
Change-Id: Ic8957296b3cfe3e6abe1218ef060614d9cdc5618
Reviewed-on: https://gerrit.iotivity.org/gerrit/9545
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jay  Sharma <jay.sharma@samsung.com>
Reviewed-by: Junghyun Oh <junghyun.oh@samsung.com>
Reviewed-by: Hun-je Yeon <hunje.yeon@samsung.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>