Fixed bug during CacheID generation in resource encapsulation.
On calling startCaching API [RCSRemoteResourceObject.h] for different resources
one after another it is generating the same CacheID.

If we call the API at 1 second delay it is working fine; but it's better to use
a more accurate random generator. OCGetRandom() function provided by CSDK
seems to be more accurate here as it seeds each byte separately.

In this patch we have replaced rand() call with OCGetRandom().
This solves the problem in our test environment and generates unique random
numbers each time which can be used as Cache IDs.

Change-Id: I52858d6b6d149ec23df39920d8a767dd08d67d69
Signed-off-by: Abhishek Pandey <abhi.siso@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9549
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Junghyun Oh <junghyun.oh@samsung.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>