Convert a data structure to class object which can be extensible for carrying more data
The current design for storing data to be delivered to enrollee and to
be arrived from Enrollee has a limitation of extensibility in case
where a vendor specific property is added to one of easy setup resources.
To overcome this situation which would happen, we convert all data structures
to class objects which can be inheritible if needed.

For example, if you add a vender specific property in DevConf resource and
want to be retrieved by getConfiguration() API, you just create NewEnrolleeConf
class inherited from EnrolleeConf class like below:
class NewEnrolleeConf : public EnrolleeConf {
   int getNewIntValue() {
       return m_DevConfRep.getValue<type>("propertykey");
} }

Change-Id: I2fa8cf53e8339c211b6b1b722bada699f5a70940
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9733
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jongho Park <jh8397.park@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>