added OCAccountManager class
As there is need for API about Cloud server on client side, I added it
on OCPlatform.
(https://gerrit.iotivity.org/gerrit/#/c/9121/)
However there are plenty of APIs to be added more and just adding them
all on OCPlatform is not fine for usability.
So I make them a class like OCResource.

Below APIs will be added after this patchset.
: searchUser, createGroup, searchGroup, deleteGroup, addGroupMember,
  searchGroupMember, deleteGroupMember

Change-Id: I45e1f08b13d821346b364d38bf212b8040205a46
Signed-off-by: Jaewook Jung <jw0213.jung@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/9835
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>