IOT-1213 Fix build failure for test build
When building with TEST=1 the build failed reporting
multiple environments were trying to build the target 'ut'

This was a hardcoded dummy target that is used to force the
unit tests to run.  The dummy target was switched from the
hardcoded 'ut' name to use the name of the unit test so two
environments don't think they are building the same target.

Change-Id: I9ac35752f0a1c7acf902bf0087a212833bf5a6a5
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/10443
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Ziran Sun <ziran.sun@samsung.com>
Reviewed-by: Jon A. Cruz <jon@joncruz.org>