Fix unittest of easy setup for TCP option build
With TCP option build, two resources per each network interface are
discovered. We have to filter out the resource with TCP flag.
Additionally, Easysetup unittest tries to discover a real provisioning
resource in a network to the test. So if there is other provisioning
resources in a network, the test would fails. To resolve the problem,
we have to filter out resources running on difference machine.

Change-Id: I130f134b051b598d6f783a6ffa9d0365fec2c172
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/10559
Reviewed-by: Heewon Park <h_w.park@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>