Fix build issue about unittest and cloud interface
When notification service unittest build with '-Wl, --as-needed' LDFLAGS,
some build machine returns build fail.(this flag is default option.)
The default of passing --as-needed can be undone by passing --no-as-needed to the linker.
So, 'Wl, --no-as-needed' flag is added in notification unittest SConscript.

CloudPublish is removed, because it is not used by sample application.

Change-Id: Ibeeb0ec60d80da1e216c43ba635e2efc0f193b7e
Signed-off-by: ch79.cho <ch79.cho@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/10627
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>