Remove consumer callback for discovered provider.
Consumer service has two type of callback within parameter as provider object.
One of callback is called when the provider discovered,
and others is called the provider changed.
But, discovered callback is imprecise called owing to provider policy.
So, discovered event is added in provider changed callback.
Two of callback is united as changed name of ProviderStateCallback
and removed discovered callback.

Change-Id: I60eebaf78aaae78a61b3772138b582df3748910c
Signed-off-by: KIM JungYong <jyong2.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/10973
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>