Fix a faulty logic in parsing a received query and in setting a URI string
When the request is sent with an empty query string, ehRequest->query is
empty string, which is interpreted as a default interface. This patch adds
a conditional statement to handle this case.
Additionally, due to https://gerrit.iotivity.org/gerrit/#/c/9111/ patch,
OCRepPayloadSetUri API is going not to be used in a simple resource like
wifi, cloudserver, and devConf resource. So, OCRepPayloadSetPropString
is used to write an URI of the resource.

Change-Id: Ie815e752007767881cef51018453e52c1a0e5863
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Signed-off-by: sy01.youn <sy01.youn@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11045
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>