[IoTivity Simulator] Updated the eclipse plugins to load the connectivity abstraction library.
Previouly CA was built as a static library but now its built as a
shared library.
Due to this, simulator eclipse plugins should explicitly load it
and update the manifest file.

Also the run-time path should be specified to help the loader to find the
dependent libraries of octbstack in resource/csdk/SConscript
for simulator to avoid the need to set the LD_LIBRARY_PATH
to run the plugins (Similar to one of our previous changes
https://gerrit.iotivity.org/gerrit/#/c/7673/).

Change-Id: I153934ecff71e9978e4726c017cdb816c96b96d6
Signed-off-by: Senthil Kumar G S <senthil.gs@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11163
Reviewed-by: Harish Marappa <h.marappa@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Tested-by: Uze Choi <uzchoi@samsung.com>