Fixed implicit-function-declaration warning for strncasecmp
Both csdk/security/src/aclresource.c and
csdk/security/src/pstatresource.c reports that there is an
implicit declaration of function 'strncasecmp' this means
the code does not see the 'strncasecmp' function from the
header.

The 'strncasecmp function exist in the strings.h header or
in string.h header. To have the funtion in the string.h
header the _GNU_SOURCE or _DEFAULT_SOURCE macro must be
defined.

Since aclresource.c explicity excludes strings.h if string.h
is found the _GNU_SOURCE macro was added.

The _GNU_SOURCE macro was used over the _DEFAULT_SOURCE
macro since there already apears to be a presidence for
the _GNU_SOURCE macro elsewhere in the code.

Change-Id: I2d29521d0f28811be2a58349a5bc8be39b17d76e
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11529
Reviewed-by: Larry Sachs <larry.j.sachs@intel.com>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>
Tested-by: Rick Bell <richard.s.bell@intel.com>