Remove OC_REST_CANCEL_OBSERVE from OCMethod.
OC_REST_CANCEL_OBSERVE was created for internal uses but
has not been used. Since it is defined as one of OCMethod,
Users attemp to call it in OCDoResrouce(...). OCDoResrouce(...)
is to discover or Perform requests on a specified resource and
OCCancel(...) should be used to cancel the operation supported in
OCDoResource(...).

Bug: https://jira.iotivity.org/browse/IOT-1240
Change-Id: Ie40d8961949f1fa3f0c2c04899c58d6ae62241af
Signed-off-by: Ziran Sun <ziran.sun@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11703
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>