[IOT-1294] Fixed build failure for WITH_RA_IBB flags
When building code with the WITH_RA_IBB=1 the build
would fail with undefined references to code found in the external
code found in extlibs/libstrophe and extlibs/wksxmppxep.

The code was already being built but was not properly linked by the
linker. This commit adds the library paths and libraries needed to
pass the build.

Change-Id: Ia9a1a6c40f6eeed174caaeb428224b97ffb6b59d
Signed-off-by: George Nash <george.nash@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11715
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Rick Bell <richard.s.bell@intel.com>
Tested-by: Rick Bell <richard.s.bell@intel.com>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>