Enable unit tests on Windows
* Add OCPlatformTest.cpp back into build process. Disabled specific tests
  that cause the delta in exception handling between Windows and Linux to
  be a problem. Replaced use of NULL with nullptr
* Add OCRepresentationEncodingTest.cpp back into build process
  only for cl.exe 14.0 (not 12.0)
* Add OCDevicePayloadCreate to octbstack_product.def
* Add OCByteStringCopy to octbstack_product.def
* Explanation:
On the Windows platform, certain exceptions are not being
caught by Gtests' EXPECT_ANY_THROW macro.
Specifically; on Windows, asynchronous structured exceptions
are not caught with C++'s try/catch exception handling. Unfortunately,
these tests raise asynchronous structured exceptions, and the result
is the unit tests crash instead of reporting a failure.

The two most common exception tests being disabled are:
1) Assigning a null pointer to a std::string
2) Indexing into an empty vector

For a specific example, where an asynchronous structured exception
is provoked, look at:
TEST(BindInterfaceToResourceTest, BindZeroResourceInterface)
inside of resource/unittests/OCPlatformTest.cpp

The tests could be refactored because both of
the operations causing the exceptions are not acceptable for the caller to
perform. Expecting anything afterward may be hazardous.
Furthermore, these tests just prove that the C++ documentation is
correct when it states that passing a null pointer to a std::string
constructor, or indexing into an empty vector is a bad idea.

"If s is a null pointer, if n == npos, or if the range specified by
[first,last) is not valid, it causes undefined behavior."
http://www.cplusplus.com/reference/string/string/string/#exceptions

"Portable programs should never call this function with an argument n
that is out of range, since this causes undefined behavior."
http://www.cplusplus.com/reference/vector/vector/operator[]/

Change-Id: If8dbdcb54d7a0fe22ca71fc8513eab413620beed
Signed-off-by: Daniel Ferguson <daniel.j.ferguson@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/11783
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>