Fix for IOT-1394.
In callbacks for GET, PUT, POST, and OBSERVE requests in
InProcClientWrapper.cpp (CPP layer of RI), the representation was parsed
only for success cases.

But when we are dealing with cloud responses,
they include a payload with error message and code even for error cases.

To catch those responses, we should ignore the status(OCStackResult)
of client response and let the error representation reach till application.

Change-Id: I103957bb79e948693807fc05e8bed461e6b2d8dd
Signed-off-by: Senthil Kumar G S <senthil.gs@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/12831
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-by: Abhishek Pandey <abhi.siso@samsung.com>
Reviewed-by: Ziran Sun <ziran.sun@samsung.com>