Change doxm.deviceuuid from R-only to RW in doxm entity handler.
CTT testing revealed a bug in 1.2 (and a matching bug in the 1.1
schema) that had marked deviceuuid as R-only.  However this is
not correct: deviceuuid must be Writeable to allow the OBT
to update with a new UUID during onboarding.  This has been
verified by SecWG and needs to be fixed in 1.2.

Patch #2: added a line to also copy deviceuuid in updateWriteableProperty() function

Change-Id: Icff6d978f5a12af021d557a5c71f25a4e9957ba2
Signed-off-by: Nathan Heldt-Sheller <nathan.heldt-sheller@intel.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14345
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Randeep Singh <randeep.s@samsung.com>