Update build scripts for changes to RD for Windows
These changes allow Windows to build correctly with
Resource Directory code. Changes from WITH_RD to RD_MODE were
introduced in early October, but the RUN.BAT file that builds
Windows was not updated, so Windows has been building with a
broken mix of RD and not RD to this point. Changes to RUN.BAT and
SCONS scripts bring Windows builds up-to-date.

Change-Id: I9ba789a11fa0ca1865822578f7da8d4a56439660
Signed-off-by: Mike Fenelon <mike.fenelon@microsoft.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13905
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>
Reviewed-by: Habib Virji <habib.virji@samsung.com>
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/13997
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Tested-by: Phil Coval <philippe.coval@osg.samsung.com>