Remove cJSON library include from resource
Removing this reduces ocbtstack from 180kb to 168kb in size.

cJSON is not used anymore in the resource, except in json2cbor tool in
the security module.

cJSON is used in coap-http and simulator and scripts has been updated
to include and build cjson. It was included mostly in sconscript.

pmutility had helper function which was not used anywhere.
smutility, psinterface, oicgroup all had function calling cjson but the
function was not used anywhere in the code. Except psinterface and
smutility there are no public API with cjson. The public API are not
called by either sample or test code. The API are obsolete and using them
will not yield any result and can crash the platform.

Change-Id: Ifbf3384536394a26cd6b168dcd15285ef9ce08a7
Signed-off-by: Habib Virji <habib.virji@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/14685
Reviewed-by: Kevin Kane <kkane@microsoft.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mike Fenelon <mike.fenelon@microsoft.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>