service: [IOT-1562] Fix for unittest with SECURED at RE/RC
[KIM JungYong]
When IoTivity was built with SECURED and TEST,
some of test case at RE and RC occurring crash.
because RE and RC has not set the precondition of security.

In this patch, added the security flow and persistence storage files at RE.
related security test case is disabled at RC.

[Philippe Coval]
Ported from 1.2-rel to master branch

Change-Id: I4094bd01cba84d395c408a56f4c57d4ebcb45b44
Signed-off-by: KIM JungYong <jyong2.kim@samsung.com>
Bug: https://jira.iotivity.org/browse/IOT-1562
Reviewed-on: https://gerrit.iotivity.org/gerrit/14837
Tested-by: Phil Coval <philippe.coval@osg.samsung.com>
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>
Signed-off-by: Phil Coval <philippe.coval@osg.samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16867
Tested-by: jenkins-iotivity <jenkins@iotivity.org>
Reviewed-by: Ziran Sun <ziran.sun@samsung.com>