Remove resource-hosting service which is deprecated.
Current resource-hosting service is deprecated because
concept of resource proxy can cover with functionality of cloud or relay server.
And this service would not handled for previous releases.
So this service can't cover for recent basic protocols.
If needs is arised for this concept, will be develop on the new service.

Change-Id: Ie14482924b7ccbebbd2512f017ad4b6039d72dd2
Signed-off-by: KIM JungYong <jyong2.kim@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15003
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15245
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Phil Coval <philippe.coval@osg.samsung.com>