Add an additional callback in security provisioning of easy setup after secure resource is discovered.
Right after a target secure resource is discovered, Mediator may need to
confirm which ownership tranfer method is used or set a pre-configured pin
number for MOT use case.
For that, it adds an additional ESResult, ES_SECURE_RESOURCE_IS_DISCOVERED,
which indicates the event. And if some user inputs are needed, user can
pass its inputs as a form of ESOwnershipTransferData object which can store
an ownership transfer method to be used and pre-configured pin to be used
for a MOT case.

Note that, related APIs for MOT was merged
: https://gerrit.iotivity.org/gerrit/#/c/14221/

Change-Id: Ic17f50693a6c4073eb5208da861506230f817f0e
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15161
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>
(cherry picked from commit 80884027f52113cdeb2673877d2a3d338630e371)
Reviewed-on: https://gerrit.iotivity.org/gerrit/15467