Changed the type of data added in ca_threadpool_pthreads list.
If u_arraylist_add() failed, pthread_join() should be called to
release its resources in ca_thread_pool_add_task().
but If the state of target thread is blocked on a condition variable,
there is no way to release its resource. (pthread_join() function
will wait for the thread to terminate.)
So, I changed the type of data added in ca_threadpool_pthreads list
to release its resource when ca_thread_pool_free() is called.

Change-Id: Ieae5d920368e46230ca312817ed1726bb08e0f2d
Signed-off-by: hyuna0213.jo <hyuna0213.jo@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15591
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Jaehong Jo <jaehong.jo@samsung.com>
Reviewed-by: Dan Mihai <Daniel.Mihai@microsoft.com>
Reviewed-by: jihwan seo <jihwan.seo@samsung.com>
Reviewed-by: Ashok Babu Channa <ashok.channa@samsung.com>