Add "accessToken" and "acceessTokenType" property in cloudserver resource.
Rather than authCode, some of IoT Cloud servers may use an access token for
enrollee to do sign-up. For example, after a mediator got issued
an auth code from account server, it can send a request to change it
to its corresponding access token which is going to deliver to Enrollee.
Additionally, AccessTokenType property is added, which indicates ,for example,
an given access token is "bearer" type token.

Change-Id: I6feb2b4f1582b5c9ee9b104b3592267247f97f99
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/15059
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Heewon Park <h_w.park@samsung.com>
Reviewed-by: Uze Choi <uzchoi@samsung.com>