Update for Multi Ownership Transfer condition
The found enrollee's owner ID indicates a same ID of mediator.
However, a list of owned devices managed in mediator's PMD db has
no element for the found enrollee. In that case, MOT will be failed.

For this, ES_OWNERSHIP_IS_NOT_SYNCHRONIZED value is returned, which
guides a user to reset a Enrollee's SVR DB file.

Change-Id: Ia5feaccccfc4cca4a0673d08cbba4d473324e37f
Signed-off-by: Parkhi <h_w.park@samsung.com>
Signed-off-by: Jihun Ha <jihun.ha@samsung.com>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16191
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Uze Choi <uzchoi@samsung.com>