IOT-1658 Added missing C Export functions to the Windows build
Added several critical methods that for some reason was missing from the
export list. Without these
methods it's impossible to perform service discovery and interaction, as
you can't work with the payloads.

You can from a Visual Studio command prompt run the following command to
get a list of the methods that are exported:
    dumpbin /EXPORTS octbstack.dll

This will confirm before/after that the methods are there or not. I've
successfully added this locally which enabled
me to invoke all of these methods from .NET Interop.

Change-Id: Ibb5857b00fa4c82228c3621ea63c4de3439649b6
Signed-off-by: Morten Nielsen <mn@iter.dk>
Reviewed-on: https://gerrit.iotivity.org/gerrit/16233
Tested-by: jenkins-iotivity <jenkins-iotivity@opendaylight.org>
Reviewed-by: Mike Fenelon <mike.fenelon@microsoft.com>
Reviewed-by: Dave Thaler <dthaler@microsoft.com>